NO-ISSUE: Move tests for ocs into ocs package
