+++ internal/operators/ocs/ocs_suite_test.go
@@ -1,4 +1,4 @@
-package ocs
+package ocs_test
 
 import (
 	"testing"
